ACL working !!!!!
Also started the user management system, admin's only can access it need to still do edit and remove actions for users but on the whole starting to look good.

Added v01.sql, This is the core db file, if some branch chnages are being made they will be made to v02 (for branch 2) and v03 for branch 3 and so on. But for now just use the v01 file. As a note the user passwords are salted and SH1 encoded so to create some accounts you need to find the salt in the login controller and concat it with your password and add etc etc